restructured folders into feature organization
